Moved snippet tests to docs package
